feat(routes): new /profile, /admin, /wiki-management landing pages
- /profile reduced to 4 user-personal sections in order:
  Knowledge Stats, Password, 4-digit passcode (disabled with
  'Available in enterprise' label), Log out.
  Removed sections (moved or deleted): MCP Connection, Prompts link,
  AI Models, User Management (vault), Data → Export buttons + data
  flow link, Danger zone, Security (split: password stays, keypair
  management moves to admin export).

- /admin landing page with inline sections at top and subroute cards
  below:
    Inline: MCP Connection (workspace endpoint + regen),
            Data (Export all data, Export keypair),
            Danger zone (Delete all data with username confirm)
    Cards:  Wiki Types, Providers + AI Models, Backfill, Spend
  In personal mode the same single user sees this page; in enterprise
  mode it will be role-gated to Admins only (gating not in this PR).

- /wiki-management landing page with two subroute cards:
    Wikis (per-wiki autoregen / regen / publish)
    People queue (auto-extracted person quarantine)
  Per-wiki Settings tab continues to live on each wiki's detail page,
  not here.
